REFACTOR: code to new SFC and applied Vue styleguide
