+++ frontend-vue/src/components/the-navigation-top.vue
@@ -1,5 +1,5 @@
 <template>
-  <nav id="main-navigation" class="app-menu">
+  <nav id="navigation-top" class="app-menu">
     <template v-for="{ path, name } in router.options.routes" :key="path">
       <router-link :to="path">{{ routeNamesTranslated[name] }}</router-link> |
     </template>
@@ -28,23 +28,23 @@ const translatorAsync = useTranslator();
 watchEffect(
   async () =>
     (routeNamesTranslated.value = await translatorAsync(
-      "components/main-navigation",
+      "components/the-navigation-top",
       store.language
     ))
 );
 </script>
 
 <style scoped>
-#main-navigation {
+#navigation-top {
   padding: 30px;
 }
 
-#main-navigation a {
+#navigation-top a {
   font-weight: bold;
   color: #2c3e50;
 }
 
-#main-navigation a.router-link-exact-active {
+#navigation-top a.router-link-exact-active {
   color: #42b983;
   filter: drop-shadow(0 2px 6px #330);
 }
