+++ frontend-vue/src/translations/en/components/the-navigation-top.js
