+++ frontend-vue/src/translations/es/components/the-navigation-top.js
