+++ frontend-vue/src/translations/et/components/the-navigation-top.js
